+++ CHANGELOG.md
@@ -40,6 +40,7 @@ Version 0.3.0
 - Fix Char initialization to " "
 - Make SimpleTestRunner to give error in exact line
 - Add reset() to Sqlca value, use reset before sql methods
+- Overflow integer not exact division 
 
 
 Version 0.2.0